requirements.txt
asyncio
random
json
os
azure-iot-device
